feat: Add User page and route
This commit adds a new page called User and updates the App component to include a route for the User page. The User page is a simple component that displays the text "halaman users". This change was made to enhance the functionality of the application by providing a new page for users to view.